+++ frontend/packages/app/src/pages/project-details/tabs/feedback/clientFeedbackView/responsesList.tsx
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies.
  */
-import type { ResponseItem } from "./types";
+import type { ResponseItem } from "../types";
 
 interface ResponsesListProps {
   responses: ResponseItem[];
